Revert "Remove `BaseFamilyTree` (#11)"
This reverts commit 0d0537c0802117e745ee431478a9cffb5481cdf4.
